Finished basic implementation of the queue data structure.
The implementation is based on maintaing the "front" index and the
number of elements "n_elem" of the queue. When it is full and a "push"
is performed, the ring buffer is doubled and linearized. The
linearization is performed by copying the data elements from "front"
to the end of the data vector to a buffer, then shifting the data
vector by the size of this buffer, then copying the buffered data to
the front of the vector.