Fixed remaining bugs with bds_serialize().
The procedure now correctly takes into account any padding between
members in a data struct. Also, the calculation of the length of the
member data segment has been simplified, where the bds_object_desc
struct now has the object length and we use it for the member data
segment length.