Fixed bug in serialize_object() for the BDS_OBJECT_PTR_DATA case which copied data beyond the bounds of the member pointer allocation region when the allocation size is not aligned to the alignment size.
The base address of serialized objects is now stored as a int64_t so
that it has a fixed byte length. The fixed byte length gives a
predictable way to ensure that external access to the base address is
done properly when the sizeof(long) is not consistent for other
runtime systems (e.g., OpenCL) depending on the architecture of the
device.